\name{cBar2-package}
\alias{cBar2-package}
\alias{cBar2}
\docType{package}
\title{
\packageTitle{cBar2}
}
\description{
\packageDescription{cBar2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cBar2}
\packageIndices{cBar2}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{cBar2}

Maintainer: \packageMaintainer{cBar2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
